+++ src/liquid_parquet/src/reader/runtime/mod.rs
@@ -8,8 +8,8 @@ mod in_memory_rg;
 mod liquid_batch_reader;
 mod liquid_predicate;
 mod liquid_stream;
+mod parquet;
 mod parquet_bridge;
-mod reader;
 mod utils;
 
 pub struct LiquidRowFilter {
